+++ R/gmus.R
@@ -7,8 +7,7 @@
 #'   error.
 #' @param family "gaussian" for linear regression, "binomial" for logistic
 #'   regression or "poisson" for Poisson regression.
-#' @return List object with intercept and coefficients at the values of lambda
-#'   and delta specified, as well as regularization parameters.
+#' @return An object of class "gmus".
 #' @references \insertRef{rosenbaum2010}{hdme}
 #'
 #'   \insertRef{sorensen2018}{hdme}
@@ -29,6 +28,7 @@
 #' fit1 <- gmus(W, y)
 #' # Draw an elbow plot to select delta
 #' plot(fit1)
+#' coef(fit1)
 #'
 #' # Now, according to the "elbow rule", choose
 #' # the final delta where the curve has an "elbow".
@@ -37,6 +37,8 @@
 #' fit2 <- gmus(W, y, delta = 0.08)
 #' # Plot the coefficients
 #' plot(fit2)
+#' coef(fit2)
+#' coef(fit2, all = TRUE)
 #'
 #' @export
 gmus <- function(W, y, lambda = NULL, delta = NULL,